Add support for out-of-core visualization.
This is based on the MmapAllocator from tgd 5.0.

Additionally, the quad tree building and handling was rewritten. The quadtree
is now always built completely on first access, and only the textures required
for rendering the current frame are uploaded.

This means the array data both for the original array and for the quad tree are
effectively stored on disk for large data sets.

Still TODO:
* progress report on opening a very large file (loading, quad tree building)
* performance improvements for building quadtree levels >= 1
* testing